% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CalGeneInfAmount}
\alias{CalGeneInfAmount}
\title{CalGeneInfAmount
Evaluate the ability of genes to respond to valid information based on the standard deviation of gene expression}
\usage{
CalGeneInfAmount(ExpMat, probs = 0.99, rescale = FALSE, plot = FALSE)
}
\arguments{
\item{ExpMat}{Normalized gene expression matrix, n genes × m samples}

\item{probs}{Quantile of non-zero gene expression values, default is 0.99}
}
\value{
Returns the informative ability of each gene, in the range 0-1. The closer to 1, the greater the ability.
}
\description{
CalGeneInfAmount
Evaluate the ability of genes to respond to valid information based on the standard deviation of gene expression
}
