% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetDivergence.R
\name{CoarseGrainingDensity}
\alias{CoarseGrainingDensity}
\title{CoarseGrainingDensity}
\usage{
CoarseGrainingDensity(rawDensity, n.coarse, n.grid = NULL)
}
\arguments{
\item{rawDensity}{a matrix,}

\item{gridPoint}{}
}
\value{

}
\description{
CoarseGrainingDensity
}
\keyword{internal}
