% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEAModel.R
\name{creatNEAModel}
\alias{creatNEAModel}
\title{creatNEAModel}
\usage{
creatNEAModel(
  counts,
  ExpData = NULL,
  do.impute = TRUE,
  n.dropGene = 3000,
  n.randNet = 20000,
  k = 10,
  Div_weight = sqrt(as.vector(matrix(1:n.coarse, nrow = n.coarse, ncol = n.coarse) *
    t(matrix(1:n.coarse, nrow = n.coarse, ncol = n.coarse)))),
  Likelihood,
  GroupLabel = NULL,
  n.grid = 60,
  n.coarse = 20,
  CoarseGrain = TRUE,
  loop.size = 3000,
  parallel.sz = 10,
  noiseSd = 0.01,
  verbose = TRUE,
  exclude.zero = FALSE,
  NoiseRemove = TRUE,
  divergence = "jsd",
  ds.method = "knn",
  h = 3,
  sdBias = 1.1,
  rb.jsd = FALSE,
  parllelModel = c("foreach", "bplapply")[1]
)
}
\arguments{
\item{counts}{a matrix represents gene count expression, with rows being genes and columns being cells}

\item{ExpData}{a matrix represents gene normalized expression, with rows being genes and columns being cells.}

\item{do.impute}{Logical,whether to perform gene expression imputation}

\item{n.dropGene}{Integer, number of cells sampled for model building (def:3000)}

\item{n.randNet}{Integer, number of cells sampled for model building (def:3000)}

\item{k}{a numerical value represents the kth (def: 10) nearest neighbor point}

\item{Div_weight}{sqrt(as.vector(matrix(1:n.coarse,nrow = n.coarse,ncol = n.coarse)*t(matrix(1:n.coarse,nrow = n.coarse,ncol = n.coarse)))),}

\item{Likelihood}{Named numeric vector,}

\item{GroupLabel}{a string vector representing the grouping information of the cells}

\item{n.grid}{An integer value representing the number of grid (def:60) used to calculate the joint probability density}

\item{n.coarse}{An integer value representing the number of grid (def:20) after coarse-graining}

\item{CoarseGrain}{Logical value, whether to coarse-grain the grid(def:TRUE).}

\item{loop.size}{An integer value representing the network size (def:3000) allocated to each compute core}

\item{parallel.sz}{An integer representing the number of parallel cores (def:1)}

\item{verbose}{Logical value, whether progression messages should be printed in the terminal.}

\item{exclude.zero}{Logical value indicating whether to exclude 0 values when calculating the joint probability density (def: FALSE)}

\item{NoiseRemove}{Logical value, indicating whether to remove noise that may lead to extremely low probability density (def:TRUE)}

\item{divergence}{Method for calculating divergence, optional jsd (Jensen-Shannon divergence) and kld (Kullback–Leibler divergence)(def:jsd)}

\item{ds.method}{Method to calculate joint probability density, optional knn and kde (def:knn)}

\item{h}{A parameter used for kernel density estimation, which only takes effect when the density estimation method is kde.}

\item{sdBias}{Numeric value, bias coefficient used to penalize low degree genes(>=1) (def:1.1)}

\item{rb.jsd}{(def:FALSE)}

\item{parllelModel}{parllelModel=c('foreach','bplapply')[1]}
}
\value{

}
\description{
creatNEAModel
}
\keyword{internal}
