% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDNSmain.R
\name{CreatScDNSobject}
\alias{CreatScDNSobject}
\title{CreatScDNSobject}
\usage{
CreatScDNSobject(
  counts,
  data = NULL,
  Network = NULL,
  GroupLabel = NULL,
  k = 10,
  n.grid = 60,
  NoiseRemove = TRUE,
  CoarseGrain = TRUE,
  n.coarse = 20,
  ds.method = c("knn", "kde")[1],
  divergence = c("jsd", "kld")[1],
  noiseSd = 0.01,
  h = 1,
  rb.jsd = F,
  Div_weight = sqrt(as.vector(matrix(1:n.coarse, nrow = n.coarse, ncol = n.coarse) *
    t(matrix(1:n.coarse, nrow = n.coarse, ncol = n.coarse)))),
  loop.size = 3000,
  parallel.sz = 3,
  verbose = TRUE,
  exclude.zero = FALSE,
  returnCDS = TRUE,
  n.dropGene = 3000,
  n.randNet = 20000,
  sdBias = 1.1,
  parllelModel = c("foreach", "bplapply")[1]
)
}
\arguments{
\item{counts}{a matrix represents gene count expression, with rows being genes and columns being cells}

\item{data}{a matrix represents gene normalized expression, with rows being genes and columns being cells.}

\item{Network}{a two-column data frame represents the biological network}

\item{GroupLabel}{a string vector representing the grouping information of the cells}

\item{k}{a numerical value represents the kth (def: 10) nearest neighbor point}

\item{n.grid}{An integer value representing the number of grid (def:60) used to calculate the joint probability density}

\item{NoiseRemove}{Logical value, indicating whether to remove noise that may lead to extremely low probability density (def:TRUE)}

\item{CoarseGrain}{Logical value, whether to coarse-grain the grid(def: TRUE).}

\item{n.coarse}{An integer value representing the number of grid (def:20) after coarse-graining}

\item{ds.method}{Method to calculate joint probability density, optional knn and kde (def:knn)}

\item{divergence}{Method for calculating divergence, optional jsd (Jensen-Shannon divergence) and kld (Kullback–Leibler divergence)(def:jsd)}

\item{h}{A parameter used for kernel density estimation, which only takes effect when the density estimation method is kde (def:1).}

\item{Div_weight}{}

\item{loop.size}{An integer value representing the network size (def:3000) allocated to each compute core}

\item{parallel.sz}{An integer representing the number of parallel cores (def:1)}

\item{verbose}{Logical value, whether progression messages should be printed in the terminal.}

\item{exclude.zero}{Logical value indicating whether to exclude 0 values when calculating the joint probability density (def: FALSE)}

\item{returnCDS}{A logical value indicating whether to return the conditional probability density}

\item{n.dropGene}{Integer, number of cells sampled for model building (def:3000)}

\item{n.randNet}{Integer, number of cells sampled for model building (def:3000)}

\item{sdBias}{Numeric value, bias coefficient used to penalize low degree genes(>=1) (def:1.1)}

\item{parllelModel}{parllelModel=c('foreach','bplapply')[1]}
}
\value{

}
\description{
creat scDNS object
}
