% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{densityCompare}
\alias{densityCompare}
\title{densityCompare
visualize the joint and conditional probability densities for a single gene-pair (an "edge") across two different biological contexts,}
\usage{
densityCompare(
  scDNSobject,
  Nodes,
  EdgeID = NULL,
  interpolate = FALSE,
  filp = FALSE,
  subEdgeID = 1,
  titleSzie = 10
)
}
\arguments{
\item{scDNSobject}{scDNS object}

\item{Nodes}{The gene(s) of interest. If two gene names are provided (e.g., c("GeneA", "GeneB")), the function plots the edge between them. If only one gene name is provided, the function attempts to plot the edge corresponding to the index specified by subEdgeID.}

\item{EdgeID}{(Optional) A pre-calculated index or logical vector specifying the exact edge row in the network to plot. If NULL (default), the function determines the edge using the Nodes parameter. Must select only one edge.}

\item{interpolate}{Specifies whether to use interpolation when rendering the density raster plots. Default is FALSE.}

\item{filp}{If TRUE, it transposes the axes (flips the X and Y variables) for the Joint Probability a}

\item{subEdgeID}{(Only used when Nodes has a length of 1) An index to select one specific edge from the list of all edges connected to the single gene provided in Nodes. Default is 1 (the first edge found).}

\item{titleSzie}{The base font size to be used for plot titles and subtitles, passed to the internal plotting function FontSize. Default is 10.}
}
\value{

}
\description{
densityCompare
visualize the joint and conditional probability densities for a single gene-pair (an "edge") across two different biological contexts,
}
