% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{densityCompare2}
\alias{densityCompare2}
\title{densityCompare2}
\usage{
densityCompare2(
  netRes,
  Nodes,
  EdgeID = NULL,
  interpolate = FALSE,
  filp = FALSE,
  subEdgeID = 1,
  titleSzie = 10
)
}
\arguments{
\item{titleSzie}{}
}
\value{

}
\description{
densityCompare2
}
\keyword{internal}
