% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{densityCompare3}
\alias{densityCompare3}
\title{Visualizing Conditional Density Distributions Across Biological Contexts}
\usage{
densityCompare3(scDNSobject, Nodes, topEdge = 10, EdgeID = NULL, ncol = 5)
}
\arguments{
\item{scDNSobject}{object of scDNS}

\item{Nodes}{A character vector specifying the gene(s) of interest.}

\item{topEdge}{The number of top-ranked edges (based on context divergence) to visualize (default is 10).}

\item{EdgeID}{An optional pre-computed logical vector to directly select the edges of interest in the network. If NULL (default), it's calculated using Nodes.}

\item{ncol}{The number of columns to use when arranging the resulting plots (default is 5).}
}
\value{
ggplot object
}
\description{
visualize and compare the conditional density distributions of gene pairs (edges) between two different biological contexts,
likely representing a control (Context A) and a stimulation or perturbed state (Context B)
}
\examples{

}
