% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{DensityPlotDF_withPoint}
\alias{DensityPlotDF_withPoint}
\title{DensityPlotDF_withPoint}
\usage{
DensityPlotDF_withPoint(
  x,
  y,
  colors = c("green", "yellow", "red", "darkred"),
  ncolor = 20,
  pt.alpha = 0.5,
  fill.alpha = 0.5,
  pt.size = 0.5,
  pt.color = "grey10",
  fill.alpha.cut = 0.15,
  showAllpoint = TRUE
)
}
\arguments{
\item{showAllpoint}{}
}
\value{

}
\description{
DensityPlotDF_withPoint
}
