% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{DetectOutlierBy3kBoxplot_UP}
\alias{DetectOutlierBy3kBoxplot_UP}
\title{DetectOutlierBy3kBoxplot_UP
Detect outliers}
\usage{
DetectOutlierBy3kBoxplot_UP(m, zero.rm = TRUE)
}
\arguments{
\item{m}{gene expression matrix, n genes × m samples}

\item{zero.rm}{logical, whether to exclude zero values}
}
\value{

}
\description{
DetectOutlierBy3kBoxplot_UP
Detect outliers
}
