% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{estimate_sigmoid_start}
\alias{estimate_sigmoid_start}
\title{estimate_sigmoid_start}
\usage{
estimate_sigmoid_start(peaks)
}
\arguments{
\item{peaks}{}
}
\value{

}
\description{
estimate_sigmoid_start
}
\keyword{internal}
