% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calZscore.R
\name{filterNetowrkFromDREMI}
\alias{filterNetowrkFromDREMI}
\title{filterNetowrkFromDREMI}
\usage{
filterNetowrkFromDREMI(scDNSob, rmP = 0.01)
}
\arguments{
\item{scDNSob}{scDNSob}

\item{rmP}{the threshold of significance(def:0.01)}
}
\value{

}
\description{
filterNetowrkFromDREMI
}
