% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calZscore.R
\name{filterNetowrkFromGAT}
\alias{filterNetowrkFromGAT}
\title{filterNetowrkFromGAT}
\usage{
filterNetowrkFromGAT(
  scDNSob,
  GAT_outNet = "./data/attention_layer1_epoch1000_wide_confidence.csv"
)
}
\arguments{
\item{scDNSob}{scDNSob}

\item{GAT_outNet}{}
}
\value{
scDNSob
}
\description{
filterNetowrkFromGAT
}
