% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{fit_conditional_curve}
\alias{fit_conditional_curve}
\title{fit_conditional_curve}
\usage{
fit_conditional_curve(peaks)
}
\arguments{
\item{peaks}{}
}
\value{

}
\description{
fit_conditional_curve
}
\keyword{internal}
