% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getDoubleDropout}
\alias{getDoubleDropout}
\title{getDoubleDropout}
\usage{
getDoubleDropout(Network, counts)
}
\arguments{
\item{counts}{gene counts expression matrix, n genes × m samples}
}
\value{

}
\description{
getDoubleDropout
}
\keyword{internal}
