% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetDivergence.R
\name{getKNNRadius}
\alias{getKNNRadius}
\title{getKNNRadius
Calculation of distance radius based on KNN method.}
\usage{
getKNNRadius(ExpDataDT, Network, GridData, k = 10)
}
\arguments{
\item{Network}{a two-column data.frame representing a biological network.}

\item{GridData}{a query data matrix. The rows are genes, and the columns are query points.}

\item{k}{a positive integer, which is the k parameter in the KNN method.}

\item{ExpData}{a data.table representing the gene expression matrix. The columns represent genes, while the rows represent samples or cells.}
}
\value{

}
\description{
getKNNRadius
Calculation of distance radius based on KNN method.
}
