% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetDivergence.R
\name{getRawDensityKNN_byRadius}
\alias{getRawDensityKNN_byRadius}
\title{getRawDensityKNN_byRadius
k/(π*r^2)}
\usage{
getRawDensityKNN_byRadius(Radius_md, k)
}
\arguments{
\item{Radius_md}{a matrix,}
}
\value{

}
\description{
getRawDensityKNN_byRadius
k/(π*r^2)
}
