% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calZscore.R
\name{getZscore}
\alias{getZscore}
\title{getZscore}
\usage{
getZscore(EdgeScore, NEAModel, Likelihood)
}
\arguments{
\item{EdgeScore}{a list or dataframe from getKLD_cKLDnetwork}

\item{NEAModel}{NEA Mdoel from creatNEAModel}

\item{Likelihood}{a vector from CalGeneInfAmount}
}
\value{

}
\description{
getZscore
}
