% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scContribution.R
\name{likelihoodFromMixEM_3dToResult_constrMean}
\alias{likelihoodFromMixEM_3dToResult_constrMean}
\title{likelihoodFromMixEM_3dToResult_constrMean
mean.constr = c(NA,0,NA)}
\usage{
likelihoodFromMixEM_3dToResult_constrMean(
  x,
  muStart = c(mean(x[x < 0]), 0, mean(x[x > 0])),
  plot = TRUE
)
}
\arguments{
\item{muStart}{}
}
\value{

}
\description{
likelihoodFromMixEM_3dToResult_constrMean
mean.constr = c(NA,0,NA)
}
\keyword{internal}
