% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Magic4MultipleData}
\alias{Magic4MultipleData}
\title{Magic4MultipleData
Imputing Gene Expression with Rmagic for Different Conditions}
\usage{
Magic4MultipleData(sob, split.by = NULL)
}
\arguments{
\item{sob}{seurat object}

\item{split.by}{group name to split data}
}
\value{
seurat object
}
\description{
Magic4MultipleData
Imputing Gene Expression with Rmagic for Different Conditions
}
