% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetDivergence.R
\name{mappingMinMaxRow}
\alias{mappingMinMaxRow}
\title{mappingMinMaxRow
maping each row of data to between minD and maxD}
\usage{
mappingMinMaxRow(ma, minD = 1, maxD = 60, rmZero = TRUE)
}
\arguments{
\item{ma}{a matrix}

\item{minD}{minimal value. The default is 1}

\item{maxD}{maximum value. The default is 60}

\item{rmZero}{Whether to exclude 0. The default is TRUE}
}
\value{
a mapped matrix
}
\description{
mappingMinMaxRow
maping each row of data to between minD and maxD
}
\examples{
mappingMinMaxRow(matrix(0:24,5),minD=1,maxD=60,rmZero = TRUE)
}
\keyword{internal}
