% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{NoAxes2}
\alias{NoAxes2}
\title{NoAxes2
NoAxes for ggplot}
\usage{
NoAxes2(
  ...,
  keep.axis.text = FALSE,
  keep.axis.title = FALSE,
  keep.ticks = FALSE
)
}
\arguments{
\item{keep.ticks}{}
}
\value{

}
\description{
NoAxes2
NoAxes for ggplot
}
