% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetDivergence.R
\name{NormalizeToOne_bygorup}
\alias{NormalizeToOne_bygorup}
\title{NormalizeToOne_bygorup
Rescale each value of the conditional density estimate by its column-maximum}
\usage{
NormalizeToOne_bygorup(ColData, label)
}
\arguments{
\item{label}{}
}
\value{

}
\description{
NormalizeToOne_bygorup
Rescale each value of the conditional density estimate by its column-maximum
}
