% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plot_diffFC_scDNS_Zscore}
\alias{plot_diffFC_scDNS_Zscore}
\title{Volcano Plot of Log2 Fold-Change versus scDNS Z-score}
\usage{
plot_diffFC_scDNS_Zscore(
  Zscores,
  scDNSob,
  sob,
  col_Zs = c("Zscores.ZsPlus"),
  col_log2Fc = "avg_log2FC",
  fc_th = 0.25,
  Zs_th = 3,
  TextSize = 2.5,
  pt.size = c(0.25, 0.5),
  group.by,
  ident.up,
  interstingGene = NULL,
  highlightGene = interstingGene,
  TopGene = NULL,
  darkOtherGene = FALSE
)
}
\arguments{
\item{Zscores}{A data frame or object containing the gene-level network stability Z-scores (calculated by scDNS)}

\item{scDNSob}{scDNS object}

\item{sob}{Seurat Object This is used to calculate Fold Change if it's not already present in Zscores.}

\item{col_Zs}{The name of the column in the Zscores data frame that contains the Z-scores (network stability metric). Default is 'Zscores.ZsPlus'}

\item{col_log2Fc}{The name of the column in the Zscores data frame that contains the avg_log2FC values. Default is 'avg_log2FC'.}

\item{fc_th}{The threshold for avg_log2FC (fold-change).  Default is 0.25.}

\item{Zs_th}{The threshold for the Z-score (network stability score). Genes with a Z-score greater than this value are considered to have significantly altered network stability. Default is 3.}

\item{TextSize}{The font size for the gene labels plotted using geom_text_repel. Default is 2.5.}

\item{pt.size}{A vector of two values specifying the point sizes: pt.size 1 for unlabeled/other genes and pt.size 2 for highlighted/labeled genes. Default is c(0.25, 0.5).}

\item{group.by}{The metadata column in the sob (Seurat object) to use for grouping and calculating fold-change (e.g., condition).}

\item{ident.up}{The identity class (group) to compare against the rest}

\item{interstingGene}{A vector of gene names to be explicitly labeled on the plot. These genes will be highlighted regardless of their score. Default is NULL}

\item{highlightGene}{A vector of gene names to be explicitly highlighted (enlarged point size) on the plot. Defaults to the value of interstingGene}

\item{TopGene}{The number of top genes based on the Zscores column (col_Zs) to be automatically labeled on the plot. Default is NULL.}

\item{darkOtherGene}{If TRUE, all genes that are not labeled or highlighted (label is NA and not a highlightGene).}
}
\value{

}
\description{
plot_diffFC_scDNS_Zscore is designed to generate a Volcano-like plot that visualizes the relationship between gene expression fold-change (Log2 FC) and a network stability score (Z-score) derived from single-cell Differential Network Stability (scDNS) analysis.
}
