% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{preFilterNet}
\alias{preFilterNet}
\title{preFilterNet}
\usage{
preFilterNet(BioNet, CountData)
}
\arguments{
\item{BioNet}{BioNet}

\item{CountData}{gene expression counts matrix ,genes × cells}
}
\value{

}
\description{
preFilterNet
}
