% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetDivergence.R
\name{removeNoiseInCDS_towDS}
\alias{removeNoiseInCDS_towDS}
\title{removeNoiseInCDS_towDS}
\usage{
removeNoiseInCDS_towDS(
  DS1,
  DS2,
  cDSList_1,
  cDSList_2,
  nx = NULL,
  min.pt = 0.5,
  min.density = 0.2864789
)
}
\arguments{
\item{cDSList_1}{data.frame(cDS_1=cDS_1,cDS_2=cDS_2)}

\item{cDSList_2}{data.frame(cDS_1=cDS_1,cDS_2=cDS_2)}

\item{nx}{}
}
\value{

}
\description{
removeNoiseInCDS_towDS
}
\examples{

}
\keyword{internal}
