% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rmNotExpressedGene}
\alias{rmNotExpressedGene}
\title{rmNotExpressedGene
Remove genes with low expression ratio in cells}
\usage{
rmNotExpressedGene(sob, min.cellN = 10)
}
\arguments{
\item{sob}{seurat object}

\item{min.cellN}{The minimum number of cells in which a gene can be expressed (def:10)}
}
\value{
sob seurat object
}
\description{
rmNotExpressedGene
Remove genes with low expression ratio in cells
}
