% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetDivergence.R
\name{rmTechnologyNoise}
\alias{rmTechnologyNoise}
\title{Title}
\usage{
rmTechnologyNoise(DensityD, min.pt = 0.7, min.density = 10/((10)^2 * pi) * 9)
}
\arguments{
\item{min.density}{}
}
\value{

}
\description{
Title
}
