% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toNetworkGeneration.R
\name{run_network_pipeline}
\alias{run_network_pipeline}
\title{run_network_pipeline}
\usage{
run_network_pipeline(
  seurat_obj,
  group_col,
  group_values = c("WT", "KO"),
  network_obj,
  random_network_obj,
  housekeeping_gmt,
  output_dir = "./output",
  run_clustering = TRUE,
  cluster_col = NULL,
  dims = 1:20,
  resolution = 0.5,
  top_n_marker = 300,
  negative_sample_size = 50000,
  save_intermediate = TRUE
)
}
\arguments{
\item{seurat_obj}{seurat_obj}

\item{save_intermediate}{}
}
\value{

}
\description{
run_network_pipeline
}
