% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scContribution.R
\name{scContribution}
\alias{scContribution}
\title{scContribution}
\usage{
scContribution(scDNSobjcet, nx = 20, rmZero = F, topGene = 100, sigGene = NULL)
}
\arguments{
\item{scDNSobjcet}{scDNSobjcet}

\item{nx}{An integer value representing the number of grid (def:20)}

\item{rmZero}{Logical value indicating whether to exclude 0 values when calculating the joint probability density (def: FALSE)}

\item{topGene}{An integer value representing The number of top genes is used to calculate the contribution of cells (def:100).The parameters will be invalid when the sigGene parameter is not NULL}

\item{sigGene}{Genes of interest are used to calculate the contribution of individual cells(def: NULL).}
}
\value{

}
\description{
scContribution
}
