% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scContribution.R
\name{scContribution_v4}
\alias{scContribution_v4}
\title{scContribution_v4}
\usage{
scContribution_v4(
  scDNSobject,
  nx = 20,
  topGene = 100,
  sigGene = NULL,
  rb = F,
  model = c("cDiv", "Div", "Plus")[3]
)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
scContribution_v4
}
