% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDNSmain.R
\name{scDNS_1_CalDivs}
\alias{scDNS_1_CalDivs}
\title{Network with divergence scores for each edge}
\usage{
scDNS_1_CalDivs(
  scDNSobject,
  k = NULL,
  n.grid = NULL,
  NoiseRemove = NULL,
  CoarseGrain = NULL,
  n.coarse = NULL,
  ds.method = NULL,
  divergence = NULL,
  h = NULL,
  noiseSd = NULL,
  loop.size = NULL,
  parallel.sz = NULL,
  verbose = NULL,
  exclude.zero = NULL,
  returnCDS = NULL,
  parllelModel = NULL
)
}
\arguments{
\item{scDNSobject}{scDNS object}

\item{k}{a numerical value represents the kth (def: 10) nearest neighbor point}

\item{n.grid}{An integer value representing the number of grid (def:60) used to calculate the joint probability density}

\item{NoiseRemove}{Logical value, indicating whether to remove noise that may lead to extremely low probability density (def:TRUE)}

\item{CoarseGrain}{Logical value, whether to coarse-grain the grid(def: TRUE).}

\item{n.coarse}{An integer value representing the number of grid (def:20) after coarse-graining}

\item{ds.method}{Method to calculate joint probability density, optional knn and kde (def:knn)}

\item{divergence}{Method for calculating divergence, optional jsd (Jensen-Shannon divergence) and kld (Kullback–Leibler divergence)(def:jsd)}

\item{h}{A parameter used for kernel density estimation, which only takes effect when the density estimation method is kde (def:1).}

\item{loop.size}{An integer value representing the network size (def:3000) allocated to each compute core}

\item{parallel.sz}{An integer representing the number of parallel cores (def:1)}

\item{verbose}{Logical value, whether progression messages should be printed in the terminal.}

\item{exclude.zero}{Logical value indicating whether to exclude 0 values when calculating the joint probability density (def: FALSE)}

\item{returnCDS}{Logical value, whether to return conditional density matrix (def: FALSE)}

\item{parllelModel}{}
}
\value{
scDNSobject
}
\description{
Computes Jensen-Shannon or Kullback-Leibler divergence between conditional density distributions of network edges (gene pairs) in two biological contexts.
}
