% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDNSmain.R
\name{scDNS_2_creatNEAModel}
\alias{scDNS_2_creatNEAModel}
\title{scDNS_2_creatNEAModel}
\usage{
scDNS_2_creatNEAModel(
  scDNSobjcet,
  n.dropGene = NULL,
  n.randNet = NULL,
  sdBias = NULL
)
}
\arguments{
\item{scDNSobjcet}{scDNSobjcet}

\item{n.dropGene}{Integer, number of cells sampled for model building (def:3000)}

\item{n.randNet}{Integer, number of cells sampled for model building (def:3000)}

\item{sdBias}{Numeric value, bias coefficient used to penalize low degree genes(>=1) (def:1.1)}
}
\value{
scDNSobjcet
}
\description{
scDNS_2_creatNEAModel
}
