% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEAModel_v2.R
\name{scDNS_2_creatNEAModel_v2}
\alias{scDNS_2_creatNEAModel_v2}
\title{scDNS_2_creatNEAModel_v2}
\usage{
scDNS_2_creatNEAModel_v2(
  scDNSobject,
  n.dropGene = NULL,
  n.randNet = NULL,
  sdBias = NULL,
  GroupLabel = NULL
)
}
\arguments{
\item{scDNSobject}{scDNS objcet}

\item{GroupLabel}{}
}
\value{

}
\description{
scDNS_2_creatNEAModel_v2
}
