% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDNSmain.R
\name{scDNS_3_GeneZscore}
\alias{scDNS_3_GeneZscore}
\title{scDNS_3_GeneZscore}
\usage{
scDNS_3_GeneZscore(scDNSobject)
}
\arguments{
\item{scDNSobject}{scDNSobject}
}
\value{
scDNSobject
}
\description{
scDNS_3_GeneZscore
}
