% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDNSmain.R
\name{scDNS_4_scContribution}
\alias{scDNS_4_scContribution}
\title{scDNS_4_scContribution}
\usage{
scDNS_4_scContribution(
  scDNSobject,
  topGene = NULL,
  Zscore_col = "combined_z",
  sigGene = NULL,
  q.th = 0.01,
  ...
)
}
\arguments{
\item{scDNSobject}{scDNSobject}

\item{topGene}{Integer value, representing the number of top significant genes(def:100).
When sigGene is not NULL, topGene is invalid.}

\item{sigGene}{A vector of strings representing the set of genes of interest (def:NULL)}

\item{q.th}{Adjusted p-value threshold}
}
\value{
scDNSobject
}
\description{
scDNS_4_scContribution
}
