% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDNSmain.R
\name{scDNS_5_cluster}
\alias{scDNS_5_cluster}
\title{scDNS_5_cluster}
\usage{
scDNS_5_cluster(
  scDNSobj,
  Sobj,
  biasToscDNS = 1,
  resolution = 0.5,
  merge.red = "pca",
  red.NewName = "scDNS"
)
}
\arguments{
\item{scDNSobj}{scDNS object}

\item{Sobj}{seurat object}

\item{biasToscDNS}{The weight of scZscore during clustering, a floating point number (default 1).
When equal to 1, scZscore and PCA's principal component 1 have the same L2 norm.}

\item{resolution}{The resolution parameter in Seurat's FindClusters function controls the number of clusters.
A higher resolution will lead to more clusters, while a lower resolution will lead to fewer clusters.}

\item{merge.red}{Single-cell transcriptomics dimensionality reduction object (default:pca) used for integration with scZScore.}

\item{red.NewName}{New scDNS dimensionality reduction name (default: scDNS)}
}
\value{
seurat object
}
\description{
scDNS_5_cluster
}
