% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEAModel.R
\name{sea_fit_fix}
\alias{sea_fit_fix}
\title{sea_fit_fix}
\usage{
sea_fit_fix(
  Dsize,
  RawScores,
  Dstd = NULL,
  winLength = 10,
  gap = 0.6,
  coarse = FALSE,
  sd_1 = NULL,
  mean_1 = NULL
)
}
\arguments{
\item{mean_1}{}
}
\value{

}
\description{
sea_fit_fix
}
\keyword{internal}
