% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SmoothOutLier_UP}
\alias{SmoothOutLier_UP}
\title{SmoothOutLier_UP}
\usage{
SmoothOutLier_UP(mt, zero.rm = TRUE, percentile = 0.99)
}
\arguments{
\item{mt}{gene expression matrix, n genes × m samples}

\item{zero.rm}{logical, whether to exclude zero values}

\item{percentile}{}
}
\value{

}
\description{
SmoothOutLier_UP
}
\keyword{internal}
