% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{theme_cowplot_i}
\alias{theme_cowplot_i}
\title{theme_cowplot}
\usage{
theme_cowplot_i(
  font_size = 14,
  font_family = "",
  line_size = 0.5,
  rel_small = 12/14,
  rel_tiny = 11/14,
  rel_large = 16/14
)
}
\arguments{
\item{rel_large}{}
}
\value{

}
\description{
theme_cowplot
}
