% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{theme_pretty}
\alias{theme_pretty}
\title{theme_pretty}
\usage{
theme_pretty(fontsize = 10, font = "Helvetica")
}
\arguments{
\item{fontsize}{font size. def:10}

\item{font}{Helvetica}
}
\value{

}
\description{
theme_pretty
}
