% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{theme_pretty_NoBox}
\alias{theme_pretty_NoBox}
\title{theme_pretty_NoBox}
\usage{
theme_pretty_NoBox(fontsize = 10, font = "Helvetica")
}
\arguments{
\item{fontsize}{font size.def:10}

\item{font}{Helvetica}
}
\value{

}
\description{
theme_pretty with no box
}
