% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{which.na}
\alias{which.na}
\title{Find indices of missing values (NA) in a vector or array.}
\usage{
which.na(x, ...)
}
\arguments{
\item{x}{An object (typically a vector or array) to check for missing values.}

\item{...}{Additional arguments passed to which(), such as arr.ind = TRUE
for array indexing.}
}
\value{
An integer vector (or array index if arr.ind = TRUE is used) giving the
positions of NA values in x.
}
\description{
This function returns the positions of NA values in the input object, by wrapping
which(is.na(x), ...) for convenience.
}
