% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEAModel_v2.R
\name{scDNS_2_creatNEAModel_v2}
\alias{scDNS_2_creatNEAModel_v2}
\title{scDNS_2_creatNEAModel_v2}
\usage{
scDNS_2_creatNEAModel_v2(
  scDNSobject,
  n.dropGene = NULL,
  n.randNet = NULL,
  repTime = 5,
  sdBias = NULL,
  GroupLabel = NULL
)
}
\arguments{
\item{scDNSobject}{scDNSobject}

\item{n.dropGene}{Integer, number of cells sampled for model building (def:3000)}

\item{n.randNet}{Integer, number of cells sampled for model building (def:20000)}

\item{repTime}{}

\item{sdBias}{Numeric value, bias coefficient used to penalize low degree genes(>=1) (def:1.1)}
}
\value{

}
\description{
Creates a null model for network enrichment analysis by random sampling of genes and networks.
}
